requirements.txt
azure-ai-agents==1.0.1
azure-ai-inference==1.0.0b9
azure-ai-projects==1.0.0b12
azure-core==1.34.0
azure-identity==1.23.0
azure-storage-blob==12.25.1
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
idna==3.10
isodate==0.7.2
markdown-it-py==3.0.0
mdurl==0.1.2
msal==1.32.3
msal-extensions==1.3.1
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
rich==14.0.0
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.0
urllib3==2.5.0
